moving code to the right place
