+++ src/test/java/com/mageddo/dnsproxyserver/dnsconfigurator/linux/LinuxDnsConfiguratorTest.java
@@ -36,90 +36,8 @@ class LinuxDnsConfiguratorTest {
     // assert
     assertEquals(
       """
-        nameserver 10.10.0.1 # dps-entry
-        """,
-      Files.readString(resolvFile)
-    );
-
-  }
-
-  @Test
-  void mustCommentExistingServerAndSetupPassedConf(@TempDir Path tmpDir) throws Exception {
-
-    // arrrange
-    final var resolvFile = tmpDir.resolve("resolv.conf");
-    Files.writeString(resolvFile, "nameserver 8.8.8.8");
-    doReturn(Collections.singletonList(resolvFile))
-      .when(this.configurator)
-      .buildConfPaths()
-    ;
-
-    // act
-    this.configurator.configure(IP.of("10.10.0.1"));
-
-    // assert
-    assertEquals(
-      """
-        # nameserver 8.8.8.8 # dps-comment
-        nameserver 10.10.0.1 # dps-entry
-        """,
-      Files.readString(resolvFile)
-    );
-
-  }
-
-
-  @Test
-  void mustUseAlreadyExistentDpsServerLine(@TempDir Path tmpDir) throws Exception {
-
-    // arrrange
-    final var resolvFile = tmpDir.resolve("resolv.conf");
-    Files.writeString(resolvFile, "nameserver 8.8.8.8\nnameserver 4.4.4.4 # dps-entry");
-    doReturn(Collections.singletonList(resolvFile))
-      .when(this.configurator)
-      .buildConfPaths()
-    ;
-
-    // act
-    this.configurator.configure(IP.of("10.10.0.1"));
-
-    // assert
-    assertEquals(
-      """
-        # nameserver 8.8.8.8 # dps-comment
-        nameserver 10.10.0.1 # dps-entry
-        """,
-      Files.readString(resolvFile)
-    );
-
-  }
-
-  @Test
-  void mustRestoreOriginalResolvconf(@TempDir Path tmpDir) throws Exception {
-
-    // arrrange
-    final var resolvFile = tmpDir.resolve("resolv.conf");
-    Files.writeString(resolvFile, """
-      # Provided by test
-      # nameserver 7.7.7.7
-      # nameserver 8.8.8.8 # dps-comment
-      nameserver 9.9.9.9 # dps-entry
-      """);
-    doReturn(Collections.singletonList(resolvFile))
-      .when(this.configurator)
-      .buildConfPaths()
-    ;
-
-    // act
-    this.configurator.restore();
-
-    // assert
-    assertEquals(
-      """
-        # Provided by test
-        # nameserver 7.7.7.7
-        nameserver 8.8.8.8
-        """,
+      nameserver 10.10.0.1 # dps-entry
+      """,
       Files.readString(resolvFile)
     );
